+++ core/src/Spirit/Parameters_EMA.cpp
@@ -52,7 +52,8 @@ try
         image->ema_parameters->n_modes = n_modes;
         image->modes.resize( n_modes );
         image->eigenvalues.resize( n_modes );
-        image->ema_parameters->n_mode_follow = std::min( image->ema_parameters->n_mode_follow, n_modes );
+        image->ema_parameters->n_mode_follow
+            = std::min( image->ema_parameters->n_mode_follow, static_cast<std::size_t>( n_modes ) );
         image->Unlock();
     }
 }
